hotfix to support multiple ui/src folders, including a robust test in the form of migrating our own code to that model, which required me to address several timing issues to ensure everything works right both logged in and logged out
