+++ modules/@apostrophecms/util/ui/src/index.js
@@ -1,0 +1,7 @@
+import util from './util.js';
+import http from './http.js';
+
+export default () => {
+  util();
+  http();
+};
