+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "apostrophe",
-  "version": "3.0.0",
+  "version": "3.0.1",
   "description": "The Apostrophe Content Management System.",
   "main": "index.js",
   "scripts": {
